Block integration test
